+++ app/src/main/java/de/baumann/browser/Activity/BrowserActivity.java
@@ -1,7 +1,6 @@
 package de.baumann.browser.Activity;
 
 import android.Manifest;
-import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.DownloadManager;
 import android.app.Notification;
@@ -127,7 +126,7 @@ import de.baumann.browser.View.SwitcherPanel;
 import static android.content.ContentValues.TAG;
 import static java.lang.String.valueOf;
 
-@SuppressWarnings({"ResultOfMethodCallIgnored", "FieldCanBeLocal"})
+@SuppressWarnings({"ResultOfMethodCallIgnored", "FieldCanBeLocal", "ApplySharedPref"})
 public class BrowserActivity extends Activity implements BrowserController, View.OnClickListener {
 
     // Menus
@@ -163,8 +162,8 @@ public class BrowserActivity extends Activity implements BrowserController, View
     private LinearLayout tv2_menu_notification;
     private LinearLayout tv2_menu_share;
 
-    LinearLayout floatButton_saveLayout;
-    LinearLayout floatButton_shareLayout;
+    private LinearLayout floatButton_saveLayout;
+    private LinearLayout floatButton_shareLayout;
 
     private View floatButton_tabView;
     private View floatButton_saveView;
@@ -461,7 +460,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
                 dialog_text.setText(helper_main.textSpannable(getString(R.string.changelog_dialog)));
                 dialog_text.setMovementMethod(LinkMovementMethod.getInstance());
 
-                FloatingActionButton fab = dialogView.findViewById(R.id.floatButton_ok);
+                ImageButton fab = dialogView.findViewById(R.id.floatButton_ok);
                 fab.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
@@ -470,7 +469,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
                     }
                 });
 
-                FloatingActionButton fab_help = dialogView.findViewById(R.id.floatButton_help);
+                ImageButton fab_help = dialogView.findViewById(R.id.floatButton_help);
                 fab_help.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
@@ -478,7 +477,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
                     }
                 });
 
-                FloatingActionButton fab_settings = dialogView.findViewById(R.id.floatButton_settings);
+                ImageButton fab_settings = dialogView.findViewById(R.id.floatButton_settings);
                 fab_settings.setOnClickListener(new View.OnClickListener() {
                     @Override
                     public void onClick(View v) {
@@ -501,7 +500,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
             }
         }, 500);
         hide_toolbar = 1;
-
     }
 
     @Override
@@ -648,7 +646,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
     }
 
     @Override
-    public synchronized void showAlbum(AlbumController controller, final boolean expand, final boolean capture) {
+    public synchronized void showAlbum(AlbumController controller, final boolean expand) {
         if (controller == null || controller == currentAlbumController) {
             switcherPanel.expanded();
             return;
@@ -743,13 +741,13 @@ public class BrowserActivity extends Activity implements BrowserController, View
 
             case R.id.tab_prev:
                 AlbumController controller = nextAlbumController(false);
-                showAlbum(controller, false, true);
+                showAlbum(controller, false);
                 viewOverflow();
                 break;
 
             case R.id.tab_next:
                 AlbumController controller2 = nextAlbumController(true);
-                showAlbum(controller2, false, true);
+                showAlbum(controller2, false);
                 viewOverflow();
                 break;
 
@@ -935,11 +933,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
                 relayoutOK.setOnTouchListener(new View.OnTouchListener() {
                     @Override
                     public boolean onTouch(View v, MotionEvent event) {
-                        if (event.getAction() == MotionEvent.ACTION_DOWN) {
-                            relayoutOK.setTextColor(ContextCompat.getColor(BrowserActivity.this, (R.color.colorAccent)));
-                        } else if (event.getAction() == MotionEvent.ACTION_UP || event.getAction() == MotionEvent.ACTION_CANCEL) {
-                            relayoutOK.setTextColor(ContextCompat.getColor(BrowserActivity.this, (R.color.light)));
-                        }
+                        relayoutOK.setTextColor(ContextCompat.getColor(BrowserActivity.this, (R.color.colorAccent)));
                         return false;
                     }
                 });
@@ -1343,7 +1337,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
 
                 if (canSwitch) {
                     AlbumController controller = nextAlbumController(left);
-                    showAlbum(controller, false, true);
+                    showAlbum(controller, false);
                 }
             }
         }));
@@ -1391,7 +1385,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
         if (currentAlbumController != null) {
             AlbumController holder;
             holder = layout;
-            showAlbum(holder, true, true);
+            showAlbum(holder, true);
             updateOmnibox();
         }
 
@@ -1408,21 +1402,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
         ImageButton open_files = layout.findViewById(R.id.open_files);
         ImageButton open_bookmark = layout.findViewById(R.id.open_bookmark);
         ImageButton open_history = layout.findViewById(R.id.open_history);
-        View home_bottomShadow = layout.findViewById(R.id.home_bottomShadow);
-
-        if (sp.getBoolean("sp_darkUI", false)){
-            open_pass.setImageResource(R.drawable.icon_key_dark);
-            open_pass.setBackgroundResource(R.drawable.custom_button_light);
-            open_newTab.setImageResource(R.drawable.icon_earth);
-            open_newTab.setBackgroundResource(R.drawable.custom_button_light);
-            open_bookmark.setImageResource(R.drawable.icon_bookmark_dark);
-            open_bookmark.setBackgroundResource(R.drawable.custom_button_light);
-            open_history.setImageResource(R.drawable.ic_action_history_dark);
-            open_history.setBackgroundResource(R.drawable.custom_button_light);
-            open_files.setImageResource(R.drawable.icon_download_dark);
-            open_files.setBackgroundResource(R.drawable.custom_button_light);
-            home_bottomShadow.setBackgroundResource(R.drawable.toolbar_drop_shadow);
-        }
 
         if (current_tab == BrowserUnit.FLAG_HOME) {
             open_newTabView.setVisibility(View.VISIBLE);
@@ -1778,7 +1757,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
                     iv.setVisibility(View.VISIBLE);
 
                     if (files_icon.matches("")) {
-                        iv.setImageResource(R.drawable.icon_arrow_up_dark);
+                        iv.setImageResource(R.drawable.file_up);
                     } else if (files_icon.matches("(.)")) {
                         iv.setImageResource(R.drawable.file_folder);
                     } else if (files_icon.matches("(.m3u8|.mp3|.wma|.midi|.wav|.aac|.aif|.amp3|.weba|.ogg)")) {
@@ -2235,8 +2214,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
         });
 
         sw_java.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-
-            @SuppressLint("ApplySharedPref")
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                 if(isChecked){
@@ -2276,8 +2253,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
         });
 
         sw_adBlock.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-
-            @SuppressLint("ApplySharedPref")
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                 if(isChecked){
@@ -2297,8 +2272,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
         }
 
         sw_image.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-
-            @SuppressLint("ApplySharedPref")
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                 if(isChecked){
@@ -2318,8 +2291,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
         }
 
         sw_cookie.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-
-            @SuppressLint("ApplySharedPref")
             @Override
             public void onCheckedChanged(CompoundButton buttonView,
                                          boolean isChecked) {
@@ -2340,8 +2311,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
         }
 
         sw_history.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-
-            @SuppressLint("ApplySharedPref")
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                 if(isChecked){
@@ -2359,8 +2328,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
         }
 
         sw_location.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-
-            @SuppressLint("ApplySharedPref")
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                 if(isChecked){
@@ -2380,8 +2347,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
         }
 
         sw_invert.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-
-            @SuppressLint("ApplySharedPref")
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                 if(isChecked){
@@ -2401,8 +2366,6 @@ public class BrowserActivity extends Activity implements BrowserController, View
         }
 
         sw_desktop.setOnCheckedChangeListener(new CompoundButton.OnCheckedChangeListener() {
-
-            @SuppressLint("ApplySharedPref")
             @Override
             public void onCheckedChanged(CompoundButton buttonView, boolean isChecked) {
                 if(isChecked){
@@ -2581,7 +2544,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
         View albumView = holder.getAlbumView();
         BrowserContainer.add(holder);
         switcherContainer.addView(albumView, LinearLayout.LayoutParams.WRAP_CONTENT, LinearLayout.LayoutParams.WRAP_CONTENT);
-        showAlbum(holder, true, true);
+        showAlbum(holder, true);
         initHomeGrid(layout);
     }
 
@@ -2615,7 +2578,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
             return;
         }
 
-        showAlbum(ninjaWebView, true, false);
+        showAlbum(ninjaWebView, true);
 
         if (url != null && !url.isEmpty()) {
             ninjaWebView.loadUrl(url);
@@ -2791,7 +2754,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
                         if (index >= BrowserContainer.size()) {
                             index = BrowserContainer.size() - 1;
                         }
-                        showAlbum(BrowserContainer.get(index), false, false);
+                        showAlbum(BrowserContainer.get(index), false);
                     }
                 });
             } else {
@@ -2801,7 +2764,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
                 if (index >= BrowserContainer.size()) {
                     index = BrowserContainer.size() - 1;
                 }
-                showAlbum(BrowserContainer.get(index), false, false);
+                showAlbum(BrowserContainer.get(index), false);
             }
         }
         showOmnibox();
@@ -3337,23 +3300,23 @@ public class BrowserActivity extends Activity implements BrowserController, View
         floatButton_shareLayout = dialogView.findViewById(R.id.floatButton_shareLayout);
 
         fab_tab = dialogView.findViewById(R.id.floatButton_tab);
-        fab_tab.setOnClickListener(this);
+        fab_tab.setOnClickListener(BrowserActivity.this);
         fab_share = dialogView.findViewById(R.id.floatButton_share);
-        fab_share.setOnClickListener(this);
+        fab_share.setOnClickListener(BrowserActivity.this);
         fab_save = dialogView.findViewById(R.id.floatButton_save);
-        fab_save.setOnClickListener(this);
+        fab_save.setOnClickListener(BrowserActivity.this);
         fab_more = dialogView.findViewById(R.id.floatButton_more);
-        fab_more.setOnClickListener(this);
+        fab_more.setOnClickListener(BrowserActivity.this);
 
         web_prev = dialogView.findViewById(R.id.web_prev);
-        web_prev.setOnClickListener(this);
+        web_prev.setOnClickListener(BrowserActivity.this);
         web_next = dialogView.findViewById(R.id.web_next);
-        web_next.setOnClickListener(this);
+        web_next.setOnClickListener(BrowserActivity.this);
 
         tab_prev = dialogView.findViewById(R.id.tab_prev);
-        tab_prev.setOnClickListener(this);
+        tab_prev.setOnClickListener(BrowserActivity.this);
         tab_next = dialogView.findViewById(R.id.tab_next);
-        tab_next.setOnClickListener(this);
+        tab_next.setOnClickListener(BrowserActivity.this);
 
         floatButton_tabView = dialogView.findViewById(R.id.floatButton_tabView);
         floatButton_saveView = dialogView.findViewById(R.id.floatButton_saveView);
@@ -3363,40 +3326,40 @@ public class BrowserActivity extends Activity implements BrowserController, View
         dialogTitle = dialogView.findViewById(R.id.dialog_title);
 
         tv_new_tabOpen = dialogView.findViewById(R.id.tv_new_tabOpen);
-        tv_new_tabOpen.setOnClickListener(this);
+        tv_new_tabOpen.setOnClickListener(BrowserActivity.this);
         tv_closeTab = dialogView.findViewById(R.id.tv_closeTab);
-        tv_closeTab.setOnClickListener(this);
+        tv_closeTab.setOnClickListener(BrowserActivity.this);
         tv_tabPreview = dialogView.findViewById(R.id.tv_tabPreview);
-        tv_tabPreview.setOnClickListener(this);
+        tv_tabPreview.setOnClickListener(BrowserActivity.this);
         tv_quit = dialogView.findViewById(R.id.tv_quit);
-        tv_quit.setOnClickListener(this);
+        tv_quit.setOnClickListener(BrowserActivity.this);
 
         tv_shareScreenshot = dialogView.findViewById(R.id.tv_shareScreenshot);
-        tv_shareScreenshot.setOnClickListener(this);
+        tv_shareScreenshot.setOnClickListener(BrowserActivity.this);
         tv_shareLink = dialogView.findViewById(R.id.tv_shareLink);
-        tv_shareLink.setOnClickListener(this);
+        tv_shareLink.setOnClickListener(BrowserActivity.this);
         tv_shareClipboard = dialogView.findViewById(R.id.tv_shareClipboard);
-        tv_shareClipboard.setOnClickListener(this);
+        tv_shareClipboard.setOnClickListener(BrowserActivity.this);
         tv_openWith = dialogView.findViewById(R.id.tv_openWith);
-        tv_openWith.setOnClickListener(this);
+        tv_openWith.setOnClickListener(BrowserActivity.this);
 
         tv_saveScreenshot = dialogView.findViewById(R.id.tv_saveScreenshot);
-        tv_saveScreenshot.setOnClickListener(this);
+        tv_saveScreenshot.setOnClickListener(BrowserActivity.this);
         tv_saveBookmark = dialogView.findViewById(R.id.tv_saveBookmark);
-        tv_saveBookmark.setOnClickListener(this);
+        tv_saveBookmark.setOnClickListener(BrowserActivity.this);
         tv_saveStart = dialogView.findViewById(R.id.tv_saveStart);
-        tv_saveStart.setOnClickListener(this);
+        tv_saveStart.setOnClickListener(BrowserActivity.this);
         tv_saveLogin = dialogView.findViewById(R.id.tv_saveLogin);
-        tv_saveLogin.setOnClickListener(this);
+        tv_saveLogin.setOnClickListener(BrowserActivity.this);
 
         tv_relayout = dialogView.findViewById(R.id.tv_relayout);
-        tv_relayout.setOnClickListener(this);
+        tv_relayout.setOnClickListener(BrowserActivity.this);
         tv_searchSite = dialogView.findViewById(R.id.tv_searchSite);
-        tv_searchSite.setOnClickListener(this);
+        tv_searchSite.setOnClickListener(BrowserActivity.this);
         tv_settings = dialogView.findViewById(R.id.tv_settings);
-        tv_settings.setOnClickListener(this);
+        tv_settings.setOnClickListener(BrowserActivity.this);
         tv_help = dialogView.findViewById(R.id.tv_help);
-        tv_help.setOnClickListener(this);
+        tv_help.setOnClickListener(BrowserActivity.this);
         tv_placeHolder = dialogView.findViewById(R.id.tv_placeholder);
         tv_placeHolder_2 = dialogView.findViewById(R.id.tv_placeholder_2);
 
@@ -3599,7 +3562,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
         TextView dialog_title = dialogView.findViewById(R.id.dialog_title);
         dialog_title.setText(getString(R.string.menu_other_help));
 
-        FloatingActionButton fab = dialogView.findViewById(R.id.floatButton_ok);
+        ImageButton fab = dialogView.findViewById(R.id.floatButton_ok);
         fab.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -3607,7 +3570,7 @@ public class BrowserActivity extends Activity implements BrowserController, View
             }
         });
 
-        FloatingActionButton fab_settings = dialogView.findViewById(R.id.floatButton_settings);
+        ImageButton fab_settings = dialogView.findViewById(R.id.floatButton_settings);
         fab_settings.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {