+++ app/src/main/java/de/baumann/browser/View/Album.java
@@ -64,7 +64,7 @@ class Album {
         albumView.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                browserController.showAlbum(albumController, false, false);
+                browserController.showAlbum(albumController, false);
             }
         });