+++ CHANGELOG.md
@@ -3,9 +3,9 @@
 - new: close all websites from notification
 - new: desktop mode in "Fast toggle dialog"
 - improved: link sharing
+- improved: light theme
 - improved: UI
 - fix: some force closes related to permission "write on storage"
-
 
 ### v 5.5
 - new: option to disable confirmation dialogs on exit (thanks: element54)